server/requirements.txt
grpcio==1.20.1
grpcio-tools==1.20.1
Pillow==6.0.0
scikit-image==0.15.0